Fix libusb locking
I was not careful enough with libusb locking (well, I used it wrong) and
this caused the application not to work. We need to unlock events before
we let libusb process them. There is no need for additional locking, all
of our calls to libusb are done after an event.